Telergam and Mattermost API moved to api package (#347)
